requirements.txt
streamlit
Pillow
streamlit-image-coordinates
opencv-python
numpy